+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * `Core`:
     * Rename of `TelegramAPIUrlsKeeper#checkWebAppLink` -> `TelegramAPIUrlsKeeper#checkWebAppData` (fix of [#591](https://github.com/InsanusMokrassar/TelegramBotAPI/issues/591))
+* `Behaviour Builder with FSM`:
+    * New typealias `StateHandlingErrorHandler`
+    * `BehaviourBuilderWithFSM` now accepts new parameter `onStateHandlingErrorHandler` which will be used in case if state has not been successfully completed
 
 ## 1.1.1